move `util` to top directory
