+++ lib/core/mapper/vrc4.dart
@@ -3,7 +3,7 @@ import 'dart:developer';
 import 'dart:typed_data';
 
 // Project imports:
-import '../util.dart';
+import '../../util.dart';
 import 'mapper.dart';
 
 // https://www.nesdev.org/wiki/VRC2_and_VRC4
